+++ frontend/src/pages/Home.css
@@ -1,0 +1,77 @@
+.home {
+	font-family: "Anek Devanagari", sans-serif;
+	color: #ffffff;
+	padding: 20px;
+}
+
+.banner {
+	background: linear-gradient(135deg, #6e8efb, #a777e3);
+	color: white;
+	padding: 40px;
+	border-radius: 8px;
+	text-align: center;
+	margin-bottom: 30px;
+}
+
+.banner h1 {
+	font-size: 2.5rem;
+	color: #ffffff;
+	margin-bottom: 10px; 
+}
+
+.banner p {
+	font-size: 1.2rem;
+	color: #ffffff;
+}
+
+.products {
+	margin-top: 20px;
+}
+
+.products h2 {
+	font-size: 1.8rem;
+	margin-bottom: 20px;
+	text-align: center;
+	color: #6e8efb;
+}
+
+.product-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+	gap: 20px;
+}
+
+.product-item {
+	font-size: 1.2rem;
+	background: #6e8efb;
+	background-image: url('/Users/Amogus/Golang-React-Ecommerce/frontend/images/orig.png'); 
+	background-size: cover; 
+	background-position: center; 
+	padding: 100px;
+	border-radius: 8px;
+	box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+	text-align: center;
+	transition: transform 0.2s;
+}
+
+.product-description {
+	position: absolute; 
+	bottom: 10px; 
+	left: 50%; 
+	transform: translateX(-50%); 
+	color: white; 
+	background: rgba(0, 0, 0, 0.5); 
+	padding: 5px 10px; 
+	border-radius: 5px; 
+	opacity: 0; 
+	transition: opacity 0.3s ease;
+	pointer-events: none;
+}
+
+.product-item:hover .product-description {
+	opacity: 1; 
+}
+
+.product-item:hover {
+	transform: translateY(-5px);
+}
