+++ frontend/src/pages/Register.css
@@ -1,0 +1,35 @@
+.register {
+	font-family: "Anek Devanagari", sans-serif;
+	color: #ffffff;
+	padding: 20px;
+    text-align: center;
+}
+
+.banner {
+	background: linear-gradient(135deg, #6e8efb, #a777e3);
+	color: 636363;
+	padding: 40px;
+	border-radius: 8px;
+	text-align: center;
+	margin-bottom: 30px;
+}
+
+input {
+    margin: 10px 0;
+    padding: 10px;
+    width: 100%;
+    max-width: 300px;
+    box-sizing: border-box;
+}
+
+button {
+    background: linear-gradient(135deg, #6e8efb, #a777e3);
+    color: white;
+    padding: 10px 20px;
+    margin-top: 10px;
+}
+
+p {
+    margin-top: 20px;
+    color: #ffffff;
+}
